# ArzTiServer/wwwroot/swagger-original.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Webservice ArzTI API",
    "description": "Kommunikation des Webservers mit den jeweiligen Rechenzentren ",
    "version": "V1"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Local server"
    },
    {
      "url": "http://arzsoftware.de/v1",
      "description": "Main (production) server"
    },
    {
      "url": "http://staging-api.arzsoftware.de",
      "description": "Internal staging server for testing"
    }
  ],
  "paths": {
...
    "/apotheke/{apoik}/rezept/status": {
      "get": {
        "tags": [
          "RezeptList"
        ],
        "summary": "Statusabfrage aller Rezepte einer Apotheke",
        "operationId": "getRezeptIdStatusList",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "query",
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "zeitraum",
            "in": "query",
            "description": "Welcher Zeitraum im Format YYYY-MM",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert eine Liste der Status aller Rezepte",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptStatus"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptStatus"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "RezeptIDList"
        ],
        "summary": "Statusänderung einer Liste von Rezepten",
        "description": "Statusänderung einer Liste von Rezepten",
        "operationId": "patchRezeptIdListStatus",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Liefert den Status einer Liste von Rezepten",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptStatus"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptStatus"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read",
              "arz_write"
            ]
          }
        ]
      }
    },
    "/apotheke/{apoik}/rezept/abholung": {
      "get": {
        "tags": [
          "RezeptList"
        ],
        "summary": "Abfrage aller neuen Rezepte einer Apotheke",
        "description": "Für die angegebene ApoIK werden alle Rezepte zurückgegeben, die noch nicht abgeholt wurden. \r\n\r\n Optional kann ein Zeitraum in der Form yyyy-mm angegeben werden und die maximale Anzahl der Datensätze, die zurückgegeben werden. \r\n\r\n Es werden solange neue Datensätze zurückgegenben, bis keine neuen mehr vorhanden sind.",
        "operationId": "getRezeptIdListAbholstatus",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "query",
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "zeitraum",
            "in": "query",
            "description": "Welcher Zeitraum im Format YYYY-MM",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert eine Liste der Status aller Rezepte",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rezept"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rezept"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "RezeptIDList"
        ],
        "summary": "Änderung des Abholstatus einer Liste von Rezepten",
        "description": "ändert Abholstatus einer Liste von Rezepten",
        "operationId": "patchRezeptIdListAbholstatus",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptId"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptId"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the newly created item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Abholstatus"
                  }
                }
              },
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Abholstatus"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      }
    },
    "/apotheke/{apoik}/rezept/pruefung": {
      "get": {
        "tags": [
          "RezeptIDList"
        ],
        "summary": "Statusabfrage aller Rezepte einer Apotheke",
        "operationId": "getRezeptIdListPruefRes",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reztyp",
            "in": "query",
            "description": ".",
            "schema": {
              "$ref": "#/components/schemas/RezeptTyp"
            }
          },
          {
            "name": "zeitraum",
            "in": "query",
            "description": "Welcher Zeitraum im Format YYYY-MM",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "style": "form",
            "explode": true,
            "description": "Status values that need to be considered for filter",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert eine Liste der Status aller Rezepte",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptPruefResult"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptPruefResult"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "RezeptIDList"
        ],
        "summary": "prüft eine Liste von Rezepten",
        "description": "Vorprüfung einer Liste von Rezepten",
        "operationId": "putRezeptIdListPruefung",
        "parameters": [
          {
            "name": "apoik",
            "in": "path",
            "required": true,
            "description": "Die Apotheken IK.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "user to add to the system",
          "content": {
            "text/plain": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptId"
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptId"
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptId"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "die Liste wurde akzeptiert"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "RezeptStatus": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "status": {
            "type": "string",
            "description": "Status des Rezepts"
          }
        }
      },
      "RezeptTyp": {
        "type": "string",
        "enum": [
          "ERezept",
          "MRezept",
          "PRezept"
        ]
      },
      "Rezept": {
        "title": "Rezept",
        "type": "object",
        "description": "base64 codierte Leistung",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "RezeptId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept ID",
        "properties": {
          "typ": {
            "$ref": "#/components/schemas/RezeptTyp"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "RezeptUId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept RUID",
        "properties": {
          "ruid": {
            "type": "string"
          }
        }
      },
      "RezeptPruefResult": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "statusinfo": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Abholstatus": {
        "type": "object",
        "required": [
          "abgeholt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "abgeholt": {
            "type": "boolean"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}